+++ buildSrc/src/main/kotlin/org/openmicroscopy/artifact/FunctionalTestPlugin.kt
@@ -1,4 +1,4 @@
-package org.openmicroscopy
+package org.openmicroscopy.artifact
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
