+++ buildSrc/src/main/kotlin/org/openmicroscopy/artifact/PublishingPlugin.kt
@@ -19,7 +19,7 @@
  *
  * ------------------------------------------------------------------------------
  */
-package org.openmicroscopy
+package org.openmicroscopy.artifact
 
 import groovy.lang.GroovyObject
 import groovy.util.Node
@@ -39,10 +39,14 @@ import org.gradle.kotlin.dsl.*
 import org.jfrog.gradle.plugin.artifactory.ArtifactoryPlugin
 import org.jfrog.gradle.plugin.artifactory.dsl.ArtifactoryPluginConvention
 import org.jfrog.gradle.plugin.artifactory.dsl.PublisherConfig
-import org.openmicroscopy.dsl.MavenPomExtensions.Companion.licenseGnu2
-import org.openmicroscopy.dsl.ProjectExtensions.Companion.camelCaseName
-import org.openmicroscopy.dsl.ProjectExtensions.Companion.getRuntimeClasspathConfiguration
-import org.openmicroscopy.dsl.ProjectExtensions.Companion.resolveProperty
+import org.openmicroscopy.artifact.dsl.MavenPomExtensions.Companion.licenseGnu2
+import org.openmicroscopy.artifact.dsl.ProjectExtensions.Companion.camelCaseName
+import org.openmicroscopy.artifact.dsl.ProjectExtensions.Companion.createArtifactoryMavenRepo
+import org.openmicroscopy.artifact.dsl.ProjectExtensions.Companion.createGitlabMavenRepo
+import org.openmicroscopy.artifact.dsl.ProjectExtensions.Companion.createStandardMavenRepo
+import org.openmicroscopy.artifact.dsl.ProjectExtensions.Companion.getRuntimeClasspathConfiguration
+import org.openmicroscopy.artifact.dsl.ProjectExtensions.Companion.resolveProperty
+import org.openmicroscopy.artifact.dsl.RepositoryHandlerExtensions.Companion.safeAdd
 import java.text.SimpleDateFormat
 import java.util.*
 
@@ -87,6 +91,12 @@ class PublishingPlugin : Plugin<Project> {
     private
     fun Project.configurePublishingExtensionSimple() {
         configure<PublishingExtension> {
+            repositories {
+                safeAdd(createArtifactoryMavenRepo())
+                safeAdd(createGitlabMavenRepo())
+                safeAdd(createStandardMavenRepo())
+            }
+
             // Publication meant for production and includes docs
             val publication = publications.create<MavenPublication>(camelCaseName()) {
                 from(components["java"])
