+++ buildSrc/src/main/kotlin/org/openmicroscopy/artifact/dsl/RepositoryHandlerExtensions.kt
@@ -1,4 +1,4 @@
-package org.openmicroscopy.dsl
+package org.openmicroscopy.artifact.dsl
 
 import org.gradle.api.Action
 import org.gradle.api.artifacts.dsl.RepositoryHandler
